Closes #1117: Replace TARA caching with storing TARA projects in a dedicated sqlite db instance
