--- iis-schemas/src/main/avro/eu/dnetlib/iis/referenceextraction/project/DocumentHashToProject.avdl
@@ -1,18 +1,0 @@
-@namespace("eu.dnetlib.iis.referenceextraction.project.schemas")
-protocol IIS{
-
-    record DocumentHashToProject {
-//      document hash value calculated as sha1 hash of 'title|abstract|text'
-        string hashValue;
-
-//      identifier of project being funding source of this document,
-//      foreign key: Project.id
-        string projectId;
-
-//		Find more details on `confidenceLevel` constraints in eu/dnetlib/iis/README.markdown file.
-        float confidenceLevel;
-
-//      text snippet surrounding the matched reference, required mostly for internal debugging and analytics
-        union { null , string } textsnippet = null;
-    }
-}
