+++ iis-wf/iis-wf-primary/src/main/resources/eu/dnetlib/iis/wf/primary/processing/oozie_app/workflow.xml
@@ -124,10 +124,6 @@
                 Set to 'a^' by default to guarantee nothing will be matched.
             </description>
         </property>
-        <property>
-            <name>tara_cache_root_dir</name>
-            <description>tara reference extraction cache root directory</description>
-        </property>
         <!-- software reference extraction related -->
         <property>
             <name>software_webcrawl_cache_location</name>
