+++ iis-wf/iis-wf-referenceextraction/src/main/resources/eu/dnetlib/iis/wf/referenceextraction/project/tara_main/oozie_app/lib/scripts/run_tara_referenceextraction.sh
