--- iis-wf/iis-wf-referenceextraction/src/test/resources/eu/dnetlib/iis/wf/referenceextraction/project/main/sampletest/data/tara_cache_output/document_hash.json
@@ -1,43 +1,0 @@
-{"hashValue":"bd056318149eadd7369cfb433a2996860b9bcd71"}
-{"hashValue":"07ea9793b2f6342fe5f3e6bf4800fdf52e0f2f36"}
-{"hashValue":"622182b68bd91d8943d6d5219e2c9f093d57c8b2"}
-{"hashValue":"22aef60cdad76c55b71b1cfa52639b81f0b39987"}
-{"hashValue":"009b8eb3d27b659eba54d9db886311e61d0530ea"}
-{"hashValue":"ca333e73b6aa32e747299156120d697c09267697"}
-{"hashValue":"c94b09b65e8d84f91b327ca00cb101bdcadb84c1"}
-{"hashValue":"3764dbc1863c149f86ecf07ffefae9f244fd095a"}
-{"hashValue":"207365e9fb42f0442fa91d939a0b6759f3cd80a2"}
-{"hashValue":"559fc3906607251e82a04989645296d13f641ff9"}
-{"hashValue":"f1ad0d7ff1737b144dc9881b6bc5db1be9b8f603"}
-{"hashValue":"4ffe1c7d29e06a8abcaaa6212dca2aa38200a470"}
-{"hashValue":"1f023b2bb30c5a8b4931f7d44623b6e0f29981e5"}
-{"hashValue":"5aa188eb1972b8773b59df8047a57503118459fd"}
-{"hashValue":"47ae144819d7cfa2c1a71285a7024a5749a72024"}
-{"hashValue":"f77b0073de9bf1d4bb125d12d7164eb457d4a923"}
-{"hashValue":"3d2fbe6fb174f6afa2537e9ce4ccdf1be121aaf9"}
-{"hashValue":"678c87f43c8e6cacf950abe5be18abb1fbd0f1ae"}
-{"hashValue":"fb522a0f623a21d88ac233aeab1e059afe12eae6"}
-{"hashValue":"4bd3264244ba34e0e87efa9e795d7f4b2c6a9aa9"}
-{"hashValue":"b13d2ff5f2e242fa124ea519f4a8e276a8ec100f"}
-{"hashValue":"aa3ea206dea5a38488203bcddd0dae7d2b6fe243"}
-{"hashValue":"1ef4001a3991149a005630033173e2a37cfc2f3b"}
-{"hashValue":"9a750b4497e7039e26496151de484ab65af5fad6"}
-{"hashValue":"5fc963afa0a40eb104b0df6b1496be391e8223ae"}
-{"hashValue":"c0b5de49a709299e95acbe6588bf79a7ec036ed9"}
-{"hashValue":"b297f7ce686e955114db74c957a635d39823230a"}
-{"hashValue":"1290566257d9b0114986a9af40d67726da986e8e"}
-{"hashValue":"2090187637da95b119e8ea7c60b5831736545bc8"}
-{"hashValue":"dff153f67c72e345e94720bc4d45df107977d63b"}
-{"hashValue":"f08602af3a14c7ef13275a4a14e2ef421c7f9fc3"}
-{"hashValue":"bb898244c60a9f4669dca97f646cb98f80f3166a"}
-{"hashValue":"f7b2143fe7628de62311ad66b10651602b933e72"}
-{"hashValue":"78d526f902d6519901b54ac7ea1a8050ff8c9a22"}
-{"hashValue":"e8e6fd1273eb639e7cf6c7f7a62dff442ba8ed1f"}
-{"hashValue":"ad0df1750e62f3a62fe657412fd5e76724c9104a"}
-{"hashValue":"a6e4142c14ebccc5b787aced63573002d2ef103e"}
-{"hashValue":"6017e71926276898a330eef4703621fc1a044297"}
-{"hashValue":"585f6941a1f4422d3a6c38688e927a1eec7e0e9d"}
-{"hashValue":"e7f226ca4b0b176ab42c4fad5d9b6dad04bd6df8"}
-{"hashValue":"bcfef12b15ca910445e76a585b94e387140435e9"}
-{"hashValue":"b40d602bc41c0fde8dc7faf531649f5f40a163a7"}
-{"hashValue":"f642e2866a150312c7ba6b57a3c3824125c357c9"}
