--- iis-wf/iis-wf-referenceextraction/src/test/resources/eu/dnetlib/iis/wf/referenceextraction/project/main/sampletest/data/tara_cache_output/document_hash_to_project.json
@@ -1,2 +1,0 @@
-{"hashValue":"f77b0073de9bf1d4bb125d12d7164eb457d4a923","projectId":"40|taraexp_____::c81e728d9d4c2f636f067f89cc14862c","confidenceLevel":0.68313,"textsnippet":null}
-{"hashValue":"1290566257d9b0114986a9af40d67726da986e8e","projectId":"40|taraexp_____::6413b29c08e6d71a9cf6c4d50d7dc6f6","confidenceLevel":0.68313,"textsnippet":null}
