--- iis-wf/iis-wf-referenceextraction/src/test/resources/eu/dnetlib/iis/wf/referenceextraction/project/main/sampletest_without_references/data/tara_cache_output/document_hash.json
@@ -1,2 +1,0 @@
-{"hashValue":"fff1bcd932f3d647cd8739c9e0c638a64690e707"}
-{"hashValue":"7a54470fa03545654a54bc66e1910bc54e49199d"}
